webadmin: Using SizeConverter.SizeUnit by DiskSizeRenderer
1.Removes DiskSizeRender.DiskSizeUnit enum and start using
SizeConverter.SizeUnit instead. Both enums are representing disk size
units but SizeUnit is used by SizeConverter utility when converting the
size from one unit to another.

2.Signature of SizeConverter convert function is changed to return
general Number and not only long as it was before. It allows to return
numbers with floating point.

+++ frontend/webadmin/modules/webadmin/src/main/java/org/ovirt/engine/ui/webadmin/widget/renderer/RebalanceFileSizeRenderer.java
@@ -13,11 +13,11 @@ public class RebalanceFileSizeRenderer<T extends Number> extends AbstractRendere
     @Override
     public String render(T size) {
         if(size.longValue() > SizeConverter.BYTES_IN_GB) {
-            return messages.rebalanceFileSizeGb(SizeConverter.convert(size.longValue(), SizeUnit.BYTES, SizeUnit.GB));
+            return messages.rebalanceFileSizeGb(SizeConverter.convert(size.longValue(), SizeUnit.BYTES, SizeUnit.GB).longValue());
         } else if(size.longValue() > SizeConverter.BYTES_IN_MB) {
-            return messages.rebalanceFileSizeMb(SizeConverter.convert(size.longValue(), SizeUnit.BYTES, SizeUnit.MB));
+            return messages.rebalanceFileSizeMb(SizeConverter.convert(size.longValue(), SizeUnit.BYTES, SizeUnit.MB).longValue());
         } else if(size.longValue() > SizeConverter.BYTES_IN_KB) {
-            return messages.rebalanceFileSizeKb(SizeConverter.convert(size.longValue(), SizeUnit.BYTES, SizeUnit.KB));
+            return messages.rebalanceFileSizeKb(SizeConverter.convert(size.longValue(), SizeUnit.BYTES, SizeUnit.KB).longValue());
         } else {
             return messages.rebalanceFileSizeBytes(size.longValue());
         }
